+++ flake.nix
@@ -172,23 +172,18 @@
                   api_url = "http://traefik:8080/api/http/routers";
                   api_auth_user = "admin";
                   api_auth_pass = "password";
-                  interval = lib.mkOption {
-                    type = lib.types.str;
-                    default = "60s";
-                    description = ''
-                      How often to poll the Traefik API for updates. Supports Go duration units (e.g., `15s`, `1m`, `1h`, or just `60` for 60 seconds).
-                    '';
-                  };
-                  process_existing = lib.mkOption {
-                    type = lib.types.bool;
-                    default = false;
-                    description = "If true, process existing routers on startup for Traefik poller.";
-                  };
-                  record_remove_on_stop = lib.mkOption {
-                    type = lib.types.bool;
-                    default = false;
-                    description = "If true, remove DNS records when router is removed for Traefik poller.";
-                  };
+                  interval = "60s";
+                  record_remove_on_stop = true;
+                  process_existing = true;
+                };
+                file = {
+                  type = "file";
+                  name = "file_example";
+                  source = "/var/lib/dns-companion/records.yaml";
+                  format = "yaml";
+                  interval = "-1";
+                  record_remove_on_stop = true;
+                  process_existing = true;
                 };
               };
               description = "Poll profiles for service/container discovery. Each key is the poller name, and the value is an attribute set of options for that poller. TLS options for Docker are nested under 'tls'.";